api/requirements.txt
asgiref==3.8.1
audioread==3.0.1
av==14.1.0
cachetools==5.5.1
celery==5.3.6
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
dataclasses-json==0.6.7
decorator==5.2.0
Django==5.1.5
django-celery-results==2.5.1
django-cors-headers==4.7.0
djangorestframework==3.15.2
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
joblib==1.4.2
kombu==5.3.4
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
marshmallow==3.26.0
msgpack==1.1.0
mypy-extensions==1.0.0
numba==0.61.0
numpy==2.1.3
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydub==0.25.1
pyparsing==3.2.1
redis==5.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
soundfile==0.13.1
soxr==0.5.0.post1
sqlparse==0.5.3
threadpoolctl==3.5.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
